+++ behave4cmd0/__all_steps__.py
@@ -9,7 +9,7 @@ from __future__ import absolute_import
 # -- IMPORT STEP-LIBRARY: behave4cmd0
 import behave4cmd0.command_steps
 import behave4cmd0.note_steps
-import behave4cmd0.log.steps
+import behave4cmd0.log_steps
 import behave4cmd0.environment_steps
 import behave4cmd0.filesystem_steps
 import behave4cmd0.workdir_steps