+++ behave4cmd0/log_steps.py
@@ -59,12 +59,22 @@ IDEA:
 
 from __future__ import absolute_import, print_function
 import logging
+import sys
 from behave import given, when, then, step
 from behave.configuration import LogLevel
 from behave.log_capture import LoggingCapture
+from behave.log_config import LoggingConfigurator
+from behave4cmd0.command_steps import (
+    step_command_output_should_contain_text,
+    step_command_output_should_not_contain_text,
+)
 from behave4cmd0.filesystem_steps import (
-    step_file_should_contain_multiline_text,
-    step_file_should_not_contain_multiline_text)
+    step_file_should_contain_text,
+    step_file_should_not_contain_text,
+)
+
+
+_PYTHON_VERSION = sys.version_info[:2]
 
 
 # -----------------------------------------------------------------------------
@@ -75,6 +85,16 @@ def make_log_record(category, level, message):
         category = None
     logger = logging.getLogger(category)
     logger.log(level, message)
+
+
+def make_formatter(format=None, datefmt=None):
+    if format and _PYTHON_VERSION >= (3, 2):
+        # -- SINCE: Python 3.2 -- logging.Formatter(..., style)
+        style = LoggingConfigurator.select_format_style(format)
+        return logging.Formatter(format, datefmt, style)
+
+    # -- OTHERWISE:
+    return logging.Formatter(format, datefmt)
 
 
 def make_log_record_output(category, level, message,
@@ -93,7 +113,8 @@ def make_log_record_output(category, level, message,
     record_data = dict(name=category, levelname=levelname, msg=message)
     record_data.update(kwargs)
     record = logging.makeLogRecord(record_data)
-    formatter = logging.Formatter(format, datefmt=datefmt)
+
+    formatter = make_formatter(format, datefmt=datefmt)
     return formatter.format(record)
 
 
@@ -116,9 +137,9 @@ class LogRecordTable(object):
         :param table:   Table w/ log-records (as :class:`behave.model.Table`)
         :param row_schema:  Log-record row schema (as dict).
         """
-        for column, value in row_schema.items():
-            if column not in table.headings:
-                table.add_column(column, default_value=value)
+        for name, value in row_schema.items():
+            if name not in table.headings:
+                table.add_column(name, default_value=value)
 
 
 # -----------------------------------------------------------------------------
@@ -216,15 +237,11 @@ def step_command_output_should_contain_log_records(context):
     """
     assert context.table, "REQUIRE: context.table"
     context.table.require_columns(["category", "level", "message"])
+
     format = getattr(context, "log_record_format", context.config.logging_format)
     for row in context.table.rows:
-        output = LogRecordTable.make_output_for_row(row, format)
-        context.execute_steps(u'''
-            Then the command output should contain:
-                """
-                {expected_output}
-                """
-            '''.format(expected_output=output))
+        expected_output = LogRecordTable.make_output_for_row(row, format)
+        step_command_output_should_contain_text(context, text=expected_output)
 
 
 @then('the command output should not contain the following log records')
@@ -244,13 +261,8 @@ def step_command_output_should_not_contain_log_records(context):
     context.table.require_columns(["category", "level", "message"])
     format = getattr(context, "log_record_format", context.config.logging_format)
     for row in context.table.rows:
-        output = LogRecordTable.make_output_for_row(row, format)
-        context.execute_steps(u'''
-            Then the command output should not contain:
-                """
-                {expected_output}
-                """
-            '''.format(expected_output=output))
+        expected_output = LogRecordTable.make_output_for_row(row, format)
+        step_command_output_should_not_contain_text(context, text=expected_output)
 
 
 @then('the command output should contain the following log record')
@@ -334,10 +346,8 @@ def step_file_should_contain_log_records(context, filename):
     context.table.require_columns(["category", "level", "message"])
     format = getattr(context, "log_record_format", context.config.logging_format)
     for row in context.table.rows:
-        output = LogRecordTable.make_output_for_row(row, format)
-        context.text = output
-        step_file_should_contain_multiline_text(context, filename)
-
+        expected_text = LogRecordTable.make_output_for_row(row, format)
+        step_file_should_contain_text(context, filename, expected_text)
 
 @then('the file "{filename}" should not contain the log records')
 @then('the file "{filename}" should not contain the log records:')
@@ -356,9 +366,8 @@ def step_file_should_not_contain_log_records(context, filename):
     context.table.require_columns(["category", "level", "message"])
     format = getattr(context, "log_record_format", context.config.logging_format)
     for row in context.table.rows:
-        output = LogRecordTable.make_output_for_row(row, format)
-        context.text = output
-        step_file_should_not_contain_multiline_text(context, filename)
+        expected_text = LogRecordTable.make_output_for_row(row, format)
+        step_file_should_not_contain_text(context, filename, expected_text)
 
 
 @step('I use "{log_record_format}" as log record format')